tests/wpt/mozilla/tests/mozilla/window_resize_not_triggered_on_load.html

<!doctype html>
<meta charset="utf-8">
<script>
<!-- This event handler needs to be registered before the first layout -->
var resize_run = false;
window.onresize = function() {
    resize_run = true;
};
</script>
<title>window.onresize should not be called when the window first loads</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id=d></div>
<script>
var t = async_test("window.onresize should not be called when the window is not being resized");
var d = document.getElementById("d");
window.getComputedStyle(d, null).getPropertyValue("width");
d.style.width = "1px";
window.onload = function() {
    t.step(function() {
        assert_true(!resize_run);
        t.done();
    });
};
</script>
